Softeng5

+++ wechat_treasure_ring/web_data/views.py
@@ -37,7 +37,7 @@ def add_guess_subject(request):
         subject.stepsB = 0
         subject.disabled = False
         subject.save()
-        return HttpResponse("success"+str(subject.id))
+        return HttpResponse("success")
     else:
         return HttpResponse("failure")
 
@@ -80,16 +80,7 @@ def save_user_bet(request):
 
 @csrf_exempt
 def get_guess_subject(request):
-    code = request.GET.get('code')
-    get_url = 'https://api.weixin.qq.com/sns/oauth2/access_token?appid=%s&secret=%s&code=%s&grant_type=authorization_code'%(AppID,AppSecret,code)
-    try:
-        f = urllib.urlopen(get_url)
-        string_json = f.read()
-        reply = json.loads(string_json)
-        openid = reply[u'openid']
-    except:
-        return HttpResponse("Invalid code")
-        #openid = "oDetGv9_dqQBv1V-0ySwyDqdLtLs"
+    openid = request.GET.get("openid")
     subjects = []
     for dbitem in GuessSubject.objects.filter(disabled=False):
         item = {
@@ -185,20 +176,8 @@ def register(request):
     user_wight = request.GET.get("weight")
     user_goal = request.GET.get("goal_step")
     user_openid = request.GET.get("openid")
-    if not ((user_sex and user_age and user_height and user_goal and user_wight and user_goal and user_openid) and (user_wight.isdigit() and user_age.isdigit() and user_height.isdigit())):
+    if not (user_goal and user_sex and user_age and user_height and user_wight and user_openid):
         return HttpResponse("failure")
-    get_url = 'https://api.weixin.qq.com/cgi-bin/token?grant_type=client_credential&appid=%s&secret=%s'%(AppID,AppSecret)
-    f = urllib.urlopen(get_url)
-    string_json = f.read()
-    reply = json.loads(string_json)
-    #print reply
-    access_token = reply[u'access_token']
-    get_url = 'https://api.weixin.qq.com/cgi-bin/user/info?access_token=%s&openid=%s&lang=zh_CN'%(access_token , user_openid)
-    f = urllib.urlopen(get_url)
-    string_json = f.read()
-    reply = json.loads(string_json)
-    nickname = reply[u'nickname']
-    headimgurl = reply[u'headimgurl']
     if RingUser.objects.filter(user_id=user_openid).exists():
         user = RingUser.objects.filter(user_id=user_openid)[0]
         user.sex = user_sex
@@ -206,8 +185,8 @@ def register(request):
         user.height=user_height
         user.weight=user_wight
         user.target=user_goal
-        user.headimgurl = headimgurl
-        user.nickname = nickname
+        user.headimgurl = ""
+        user.nickname = "e"
         user.save()
     else:
         user_new = RingUser(
@@ -219,8 +198,8 @@ def register(request):
             target=user_goal,
             last_record=0,
             steps_totalused=0,
-            headimgurl=headimgurl,
-            nickname=nickname
+            headimgurl="",
+            nickname=""
         )
         user_new.save()
         user_bird = BirdUser(
@@ -271,8 +250,10 @@ def get_partial_ranklist(openid, objects, type):
         indexlist = [0, 1, 2, 3]
     elif n == l - 1 or n == l - 2:
         indexlist = [l - 4, l - 3, l - 2, l - 1]
-    else:
+    elif n > 0 and n < l - 1:
         indexlist = [n - 1, n, n + 1, n + 2]
+    else:
+        return []
     if l <= 4:
         indexlist = range(l)
     for i in indexlist:
@@ -354,60 +335,7 @@ def end_game(request):
     return HttpResponse(json.dumps({"result": "success"}))
 
 
-@csrf_exempt
-def game_rank(request):
-    game = request.GET.get('game')
-    start = request.GET.get('start')
-    end = request.GET.get('end')
-    openid = request.GET.get('openid')
-    attention = RecordAttention.objects.filter(source_user_id=openid)
-    attention_list = []
-    for val in attention:
-        attention_list.append(val.target_user_id)
-    if game == "bird":
-        results_today = BirdUser.objects.all().order_by('-score_today')[start:end]
-        results_total = BirdUser.objects.all().order_by('-score_total')[start:end]
-    elif game == "jump":
-        results_today = JumpUser.objects.all().order_by('-score_today')[start:end]
-        results_total = JumpUser.objects.all().order_by('-score_total')[start:end]
-    ranklist_today = []
-    ranklist_total = []
-    for item in results_today:
-        itemuser = RingUser.objects.get(user_id=item.openid)
-        if item.openid in attention_list:
-            is_attention = True
-        else:
-            is_attention = False
-        rankitem = {
-            "openid": item.openid,
-            "nickname": itemuser.nickname,
-            "headimgurl": itemuser.headimgurl,
-            "score": item.score_today,
-            "is_attention": is_attention
-        }
-        ranklist_today.append(rankitem)
-    for item in results_total:
-        itemuser = RingUser.objects.get(user_id=item.openid)
-        if item.openid in attention_list:
-            is_attention = True
-        else:
-            is_attention = False
-        rankitem = {
-            "openid": item.openid,
-            "nickname": itemuser.nickname,
-            "headimgurl": itemuser.headimgurl,
-            "score": item.score_total,
-            "is_attention": is_attention
-        }
-        ranklist_total.append(rankitem)
-    result = {
-        "today":ranklist_today,
-        "total":ranklist_total
-    }
-    return HttpResponse(json.dumps(result))
-
-
-def get_rankList(results, attention_list):
+def get_rankList(results, attention_list, type):
     ranklist = []
     for item in results:
         itemuser = RingUser.objects.get(user_id=item.openid)
@@ -415,29 +343,32 @@ def get_rankList(results, attention_list):
             is_attention = True
         else:
             is_attention = False
-        rankitem = {
-            "openid": item.openid,
-            "nickname": itemuser.nickname,
-            "headimgurl": itemuser.headimgurl,
-            "score": item.score_today,
-            "is_attention": is_attention
-        }
+        if type == 0:
+            rankitem = {
+                "openid": item.openid,
+                "nickname": itemuser.nickname,
+                "headimgurl": itemuser.headimgurl,
+                "score": item.score_today,
+                "is_attention": is_attention
+            }
+        elif type == 1:
+            rankitem = {
+                "openid": item.openid,
+                "nickname": itemuser.nickname,
+                "headimgurl": itemuser.headimgurl,
+                "score": item.score_total,
+                "is_attention": is_attention
+            }
+
         ranklist.append(rankitem)
     return ranklist
 
 
 def get_game_rank(request):
-    code = request.GET.get('code')
-    get_url = 'https://api.weixin.qq.com/sns/oauth2/access_token?appid=%s&secret=%s&code=%s&grant_type=authorization_code'%(AppID,AppSecret,code)
-    try:
-        f = urllib.urlopen(get_url)
-        string_json = f.read()
-        reply = json.loads(string_json)
-        openid = reply[u'openid']
-    except:
-        return HttpResponse("Invalid code")
-        #openid = "oDetGv9_dqQBv1V-0ySwyDqdLtLs"
-
+    openid = request.GET.get('openid')
+    ifuser = RingUser.objects.filter(user_id=openid)
+    if not ifuser:
+        return HttpResponse("User Not Exists")
     attention = RecordAttention.objects.filter(source_user_id=openid)
     attention_list = []
     for val in attention:
@@ -447,10 +378,10 @@ def get_game_rank(request):
     results_today_jump = JumpUser.objects.all().order_by('-score_today')[0:100]
     results_total_jump = JumpUser.objects.all().order_by('-score_total')[0:100]
 
-    ranklist_today_bird = get_rankList(results_today_bird, attention_list)
-    ranklist_total_bird = get_rankList(results_total_bird, attention_list)
-    ranklist_today_jump = get_rankList(results_today_jump, attention_list)
-    ranklist_total_jump = get_rankList(results_total_jump, attention_list)
+    ranklist_today_bird = get_rankList(results_today_bird, attention_list, 0)
+    ranklist_total_bird = get_rankList(results_total_bird, attention_list, 1)
+    ranklist_today_jump = get_rankList(results_today_jump, attention_list, 0)
+    ranklist_total_jump = get_rankList(results_total_jump, attention_list, 1)
     result = {
         "openid":openid,
         "today_bird":ranklist_today_bird,
@@ -462,16 +393,7 @@ def get_game_rank(request):
 
 
 def get_sleepdata(request):
-    code = request.GET.get('code')
-    get_url = 'https://api.weixin.qq.com/sns/oauth2/access_token?appid=%s&secret=%s&code=%s&grant_type=authorization_code'%(AppID,AppSecret,code)
-    try:
-        f = urllib.urlopen(get_url)
-        string_json = f.read()
-        reply = json.loads(string_json)
-        openid = reply[u'openid']
-    except:
-        return HttpResponse("Invalid code")
-    #return HttpResponse(json.dumps({'isnull':True}))
+    openid = request.GET.get("openid")
     data = {}
     if not RingUser.objects.filter(user_id=openid).exists():
         data['isnull'] = True
@@ -490,16 +412,7 @@ def get_sleepdata(request):
 
 
 def get_sportsdata(request):
-    code = request.GET.get('code')
-    get_url = 'https://api.weixin.qq.com/sns/oauth2/access_token?appid=%s&secret=%s&code=%s&grant_type=authorization_code'%(AppID,AppSecret,code)
-    try:
-        f = urllib.urlopen(get_url)
-        string_json = f.read()
-        reply = json.loads(string_json)
-        openid = reply[u'openid']
-    except:
-        return HttpResponse("Invalid code")
-    #return HttpResponse(json.dumps({'isnull':True}))
+    openid = request.GET.get('openid')
     data = {}
     if not RingUser.objects.filter(user_id=openid).exists():
         data['isnull'] = True
@@ -517,15 +430,7 @@ def get_sportsdata(request):
 
 
 def get_time_line_data(request):
-    code = request.GET.get('code')
-    get_url = 'https://api.weixin.qq.com/sns/oauth2/access_token?appid=%s&secret=%s&code=%s&grant_type=authorization_code'%(AppID,AppSecret,code)
-    try:
-        f = urllib.urlopen(get_url)
-        string_json = f.read()
-        reply = json.loads(string_json)
-        openid = reply[u'openid']
-    except:
-        return HttpResponse("Invalid code")
+    openid = request.GET.get('openid')
     data = {}
     if not RingUser.objects.filter(user_id=openid).exists():
         data['isnull'] = True
@@ -564,7 +469,9 @@ def get_time_line_data(request):
 
 def cancel_follow(request, message):
     source_id, target_id = message.split('@')
-    RecordAttention.objects.get(source_user_id=source_id, target_user_id=target_id).delete()
+    data = RecordAttention.objects.filter(source_user_id=source_id, target_user_id=target_id)
+    if data:
+        RecordAttention.objects.get(source_user_id=source_id, target_user_id=target_id).delete()
     return HttpResponse('OK')
